requirements.txt
flask
resend
requests
python-dotenv
gunicorn
